[style]컴포넌트 분리_yr[220510]
