+++ src/components/search/index.js
@@ -1,5 +1,6 @@
 import SearchBar from './SearchBar';
 import SearchHistory from './SearchHistory'
+import MapOfficeResult from './MapOfficeResult';
 
 
-export { SearchBar, SearchHistory };
+export { SearchBar, SearchHistory, MapOfficeResult };
